+++ output/decompilation/ida/metal_crusher/_ZN13metal_crusher3fun19reanimate_gui_shell17h0e56d731a5a95ba2E.c
@@ -5,7 +5,7 @@ __int64 __fastcall metal_crusher::fun::reanimate_gui_shell()
 
   v2 = v0;
   if ( (unsigned __int8)metal_crusher::fun::is_kde_plasma() )
-    return metal_crusher::fun::execute_shell(aPlasmashell_0, 11LL);
+    return metal_crusher::fun::execute_shell((__int64)aPlasmashell_0, 11LL);
   else
     return v2;
 }
